Refactor Tile and GridManager for improved visibility logic
Changed Tile from class to struct and refactored visibility logic to use toggling methods instead of setters with parameters. Updated GridManager to use new Tile constructor and visibility logic. Adjusted tile spacing and background initialization in GameState. Minor code cleanup and type consistency improvements in BackgroundManager.

+++ src/BackgroundManager.h
@@ -14,8 +14,8 @@ namespace sf {
 class BackgroundManager
 {
 private:
-	int m_loopWidth{ 0 };
-	int m_scrollSpeed{ 0 };
+	float m_loopWidth{ 0 };
+	float m_scrollSpeed{ 0 };
 	float m_scrollOffset{ 0 };
 	sf::IntRect m_drawRect;
 	std::unique_ptr<sf::Texture>m_backgroundTexture;
